Reestruturação do domínio de startups e remoção de camadas redundantes
