--- functions/src/domains/startups/services/getStartupsService.ts
@@ -1,5 +1,0 @@
-import {fetchAllStartups} from "../repositories/startupRepository";
-
-export async function getStartupsData(): Promise<Array<Record<string, unknown>>> {
-  return fetchAllStartups();
-}
