+++ functions/src/domains/startups/shared/auth.ts
