+++ functions/src/domains/startups/shared/constants.ts
