+++ functions/src/domains/startups/shared/firebase.ts
