+++ functions/src/domains/startups/shared/validation.ts
