+++ functions/src/domains/startups/types/index.ts
