--- functions/src/startups/index.ts
@@ -1,6 +1,0 @@
-// Desenvolvido por Miguel Castro
-
-export { getStartupById } from "./handlers/getStartupById";
-export { getStartups } from "./handlers/getStartups";
-export { createStartupQuestion } from "./handlers/createStartupQuestion";
-export { getStartupQuestions } from "./handlers/getStartupQuestions";
